refactor: move data type AST and parsing to specific project
This way, huge files can be avoided
