+++ Cargo.toml
@@ -7,8 +7,10 @@ edition = "2021"
 
 [dependencies]
 nom = "7"
-thiserror = "1.0"
+thiserror = "1"
 
 [dev-dependencies]
-pretty_assertions = "1.3"
-test-case = "2.2"
+pretty_assertions = "1"
+test-case = "2"
+derive_builder = "0.11"
+spectral = "0.6"
